+++ SimTKcommon/include/SimTKcommon/internal/Exception.h
@@ -9,7 +9,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org/home/simbody.  *
  *                                                                            *
- * Portions copyright (c) 2005-12 Stanford University and the Authors.        *
+ * Portions copyright (c) 2005-15 Stanford University and the Authors.        *
  * Authors: Michael Sherman                                                   *
  * Contributors:                                                              *
  *                                                                            *
@@ -24,11 +24,6 @@
  * limitations under the License.                                             *
  * -------------------------------------------------------------------------- */
 
-// Keeps MS VC++ 8 quiet about sprintf, strcpy, etc.
-#ifdef _MSC_VER
-#pragma warning(disable:4996)
-#endif
-
 #include "SimTKcommon/internal/common.h"
 
 #include <string>
@@ -40,6 +35,12 @@
 namespace SimTK {
 
 namespace Exception {
+
+    // Keeps MS VC++ quiet about sprintf, strcpy, etc.
+#ifdef _MSC_VER
+#pragma warning(push)
+#pragma warning(disable:4996)
+#endif
     
 // SimTK::Exception::Base    
 class Base : public std::exception {
@@ -302,6 +303,10 @@ public:
     }    
     virtual ~Cant() throw() { }
 };
+
+#ifdef _MSC_VER
+#pragma warning(pop)
+#endif
 
 } // namespace Exception
 } // namespace SimTK